Implemented research of data in db
Inspired to the video at the link below:
It is now possible to perform a research of the data inside the database.
https://youtu.be/2XuxFi85GTw?si=UfbrSYoaP-Szt9W6

+++ testsitecodedbyme/includes/submit_delete.php
@@ -1,10 +1,12 @@
 <?php
 require_once('../config/db.php');
 
-$id = $_POST['id'];
+$id = $_POST['id_to_delete'];
 
-$query = "DELETE FROM `wp_orderinfo` WHERE `wp_orderinfo`.`order_ID` = $id;";
-mysqli_query($con, $query);
+// Check filed is not empty
+if (!empty($id)) {
+    $query = "DELETE FROM `wp_orderinfo` WHERE `wp_orderinfo`.`order_ID` = $id;";
+    mysqli_query($con, $query);
+}
 
-header ("Location: ../style/index.php?submit_delete=success");
-?>
+header("Location: ../style/index.php?submit_delete=success");
